ptlrvp_npcx: Add lid open as PSL wake source
Add the lid open signal as a PSL wake source as this is expected
behavior.  Note that this signal is not powered in G3/Z5/Z2 (Z1 and
above only) on the PTL RVP, so it is not fully functional. This will be
fixed on a future board.

BUG=b:352318293,b:358178578
TEST=zmake build ptlrvp_npcx

Change-Id: Icf5b2914e34f6035e3fbe5743019a1c732fe0698
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6172768
Reviewed-by: Keith Short <keithshort@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Auto-Submit: caveh jalali <caveh@chromium.org>
Tested-by: caveh jalali <caveh@chromium.org>
Reviewed-by: Brandon Breitenstein <brandon.breitenstein@intel.com>
Commit-Queue: caveh jalali <caveh@chromium.org>