trulo: Configure ec_wp_odl as active low
This patch configures the ec_wp_odl GPIO with GPIO_ACTIVE_LOW, to
ensure the EC will interpret a low voltage on GPIOA pin 1 as the active
state for disabling the write protection.

The changes are generated by the arbitrage tool after the write
protection configuration is overridden in the EC GPIO arb spreadsheet.

BUG=b:386741705
TEST=zmake build trulo
TEST=Verified that EC WP GPIO and GSC WP are in sync.

Change-Id: I7e0980f36dee850bb6a7d36a38fd54bb2192a3e4
Signed-off-by: Dinesh Gehlot <digehlot@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6168868
Reviewed-by: Kapil Porwal <kapilporwal@chromium.org>
Reviewed-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Subrata Banik <subratabanik@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>