util/itecomdbgr: Fixed mode entry problem
In the past,when flashing, there was a situation
where it was judged as Invalid EFLASH Type because
there was a problem with the value read back, and
the flash mode would be re-entered several times.

Currently,this patch has corrected this judgment method.


BUG=b:328620169
TEST=Run the flash sequence over 900 times


Change-Id: Ie40b08441f2822d91283e5f332fb56bfb872df78
Signed-off-by: Donald Huang <donald.huang@ite.com.tw>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6171571
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: caveh jalali <caveh@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Donald Huang <donald.huang@ite.corp-partner.google.com>
Reviewed-by: caveh jalali <caveh@chromium.org>