pdc: Allow PDC subsystem initialization to fail
If any PDC drivers fail to initialize, pdc_power_mgmt will panic and
reboot the EC. This is not necessary. The EC and OS are capable of
booting without PD support present, so don't consider this a fatal
error.

Relaxing this requirement allows for future runtime selection of drivers
or adapting to a variable number of installed USB-C ports, from a single
EC image.

If a PDC is not ready by the time of pdc_power_mgmt initialization, that
pdc_power_mgmt driver will fail to init and enter the "Disabled" state
permanently to lock out any public API calls. This is non-fatal and
booting will continue.

The PPM will continue to initialize if only a portion of the PDC drivers
are active to allow for variable port numbers. However, if none of the
PDC drivers are ready (due to, e.g., missing configuration) it will
non-fatally fail to initialize.

Update the PPM test to wait for the PDC driver to finish its
initialization before sending UCSI commands (fix a test race condition).

BUG=b:372910804
TEST=./twister -T zephyr/test/pdc

Change-Id: I94f1fe21e9dba846d88bb98f49f76aa2ae106f0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6378603
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>