common: Fix repeated printing from panic_puts()
The panic_...() set of functions will output synchronously to the
console UART, and also queue data into the USB console buffers.
Typically, the latter does not matter, as the data is not going to make
it out of the USB buffers when the system is crashing and will reset in
a moment.

However, the "panicinfo" console command can be executed after reboot,
and it uses the same shared routines for dumping register contents as
the panic handler.

Using "panicinfo" from the USB console (rather than serial console)
revealed a funny bug, it seems that invoking
`panic_puts("abracadabra\n");` would output a repeated list of suffixes
as shown below, because `usb_puts()` prints a null-terminated string,
and not a single character, as was assumed by whoever put it inside the
loop over `outstr`.

```
abracadabra
bracadabra
racadabra
acadabra
cadabra
adabra
dabra
abra
bra
ra
a

```

BUG=none
TEST=Observe non-repeated output from panicinfo

Change-Id: I3ea7626e5759038fd8cefdd3243ec29d22665b9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6219795
Tested-by: Jes Klinke <jbk@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Jes Klinke <jbk@chromium.org>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>