util: Add script to check Zephyr code for EC symbols
In order to make it easier to upstream Zephyr code in the future, we
want to avoid using EC symbols in the zephyr directory. This script adds
a way to check for specific EC symbols and suggest the equivalent Zephyr
version to use instead.

BUG=b:393593789
TEST=./util/check_zephyr_for_ec_symbols.py zephyr/shim/src/hwtimer.c
TEST=./util/check_zephyr_for_ec_symbols.py chip/host/system.c
TEST=./util/check_zephyr_for_ec_symbols.py --commit pre-submit

Change-Id: I4f9bfbdacd5de341655ce838bcd213a03d9ea52f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6217547
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>