chip/stm32: Declare SAI addresses for STM32L4/5
Add declaration of the base address of "serial audio interface"
peripherals.

Though unlikely that we want to process digital audio on ECs, it could
be useful on the HyperDebug device to use the SAI peripherals for
outputting an audio "master clock".

In particular, the audio clock can come from the P-channel of PLLSAI1/2,
which have wider selection of divisors compared to the R-channel of the
main PLL, which would otherwise be the only option if using e.g. timers
for outputting a signal with a particular frequency.

BUG=b:384755415
TEST=make BOARD=hyperdebug

Change-Id: I1f5bb3b45abd7a218377229f69a0bed59c54ab4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6244271
Tested-by: Jes Klinke <jbk@chromium.org>
Commit-Queue: Jes Klinke <jbk@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>