npcx_monitor: Add a check for modifications
If the npcx_monitor binary changes, it needs to be updated in the servo
docker container.

Add a CMake script to check the sha256 of the output binary and fail the
build when detecting a mismatch.  The error message points the user to
check the README.md file for instructions on how to continue.

BUG=none
TEST=zmake build npcx_monitor - passes
TEST=zmake build npcx_monitor - fails on code change

Change-Id: I0cb3ce11e9fd497049c3a9ca696904bfc73941fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6245179
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>