zmake: Don't use the checkout path directly
Don't use the checkout path directly as the checkout path is not
reachable in bazel. Use the module_paths[] instead, which are the
CROS_WORKON_DESTDIR when building the package or the checkout path
when building inside chroot.

BUG=b:393355553
TEST=Build the zephyr-npcx-monitor package for fizz-labstation and
the chroot environment.
TEST=Build the chromeos-zephyr package for brox.
TEST=Build the EC image inside chroot using zmake.
TEST=Manually block the access to /mnt/host/source/src/platform,
/mnt/host/source/src/third_party, and /mnt/host/source/.repo during
the zmake execution and still pass.
TEST=zephyr/zmake/run_tests.sh

Change-Id: I55c0b2fdd58e7a2ec7e35ac42ecfd441ef421ab5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6256205
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>