chip/stm32: Add register declarations for SAI
Add declarations of STM32 Serial Audio Interface (SAI).

A few STM32 chip have support for serial audio.  We do not plan on
sending actual audio samples, but the "master clock out" feature of the
SAI blocks could come in handy, allowing the HyperDebug device to
generate a 48MHz or 96Mhz reference clock for the chip being tested.

The SAI peripherals are among the few which can be clocked from the
second and third PLL clock generator in the STM32L5, which is
particularly handy, as it allows the audio "master clock" output to not
have any particular relation to the core clock frequency of HyperDebug.

BUG=b:384755415
TEST=make BOARD=hyperdebug

Change-Id: I907f7403ddac225a1ef4032e5959e46cdf4bc3f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6307305
Tested-by: Jes Klinke <jbk@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Jes Klinke <jbk@chromium.org>
Reviewed-by: Eric Yilun Lin <yllin@google.com>