fatcat_npcx9m7f: fatcat_it82002aw: Track ACOK signal
The EC needs to track the state of the ACOK signal. This is generated by
the charger chip to indicate if an A/C charger is present (either USB
PD, or barrel jack). The hibernate implementation only considers
hibernation when the system is not on A/C.

BUG=b:391953691
TEST=npcx: "hibernate" on EC console works, power button wakes up system
	ite: EC no longer complains about "AC on, skip hibernate" but
	does not hibernate.

Change-Id: I704b5372733a75ab9fd0ed0c230fd06ec2045287
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6320413
Reviewed-by: Keith Short <keithshort@chromium.org>
Auto-Submit: caveh jalali <caveh@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: caveh jalali <caveh@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>