chip/mt_scp: Fix loop unroll to fix compilation with clang
CSR_MPU_L(i) is not a constant value, so this fails to compile. For gcc,
it was worked around by forcing a loop unroll, but this doesn't work
with clang. Instead, we change the code to unroll the loop with a macro.

BUG=b:172886808, b:407560601, b:172020503
TEST=make V=1 CROSS_COMPILE_CC_NAME=clang BOARD=geralt_scp_core1 -j
TEST=./util/compare_build.sh -b all

Change-Id: Ib74c6db263af91191b3dcf2496388a8e1412df46
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6317341
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>
Tested-by: Tom Hughes <tomhughes@chromium.org>