ppm: Move the ePPM main from init to EC main
The ePPM's main function creates a thread and waits for it to be
initialized and ready for use. This thread starts running in
INIT_APPLICATION, unlike other threads, that starts later in the
EC app main function. However, the ePPM thread doesn't need to
start that early, as no other threads depend on it.

This CL moves its execution to the EC app's main function.

The PPM must be ready by the time the OPM initializes. The OPM
is the kernel UCSI driver. Booting into the kernel takes far
longer (>1s) than the PPM thread's creation and startup. So the
PPM should be ready to use.

BUG=b:401543993
TEST=Booted Brox and login; tested power sinking and power
sourcing on the both ports.
TEST=./twister -civ -T zephyr/test/pdc

Change-Id: I38bfe11c967f1f733f6b11d4ad416bc419efbb8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6353879
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>