pdc: Move GET_CABLE_PROPERTY to end of PDM state machines
This CL moves the substates used for sending GET_CABLE_PROPERTY to the
substate just prior to SNK/SRC_ATTACHED states. The motivation for
this change is that if this command is sent prior the PD contract
being established, the PDC can reject this command. Moving the
substate ensures that a PD contract is established and active when
this command is sent to the PDC.

BUG=b:371053377
TEST=./twister -c -T zephyr/test/pdc

Change-Id: I5a58b2f27d2c708e75513a9b9211913256b9b032
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6386985
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Scott Collyer <scollyer@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Scott Collyer <scollyer@chromium.org>