chip/it83xx: Disable RAM code when building with clang
When building with clang, the code does not fit into the RAM code
section. Since the RAM code is a performance optimization and we're not
actually planning to ship the code compiled by clang for the EC boards,
just disable the RAM code when building with clang.

BUG=b:409069694, BUG=b:407560601, b:172020503
TEST=./util/compare_build.sh -b all

Change-Id: I9ce4a576a1e6a9254c42e6ddec2161a48cf451a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6437151
Reviewed-by: Dino Li <Dino.Li@ite.com.tw>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>
Tested-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Eric Yilun Lin <yllin@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>