fatcat_npcx9m7f: Add EC_PDB_CTRL GPIO
EC_PDB_CTRL isn't connected on the PTL RVP, but define it anyway for
consistency with future platforms. This is intended to be the audio amp
mute/disable function. Initialize the signal to low so the amp can be
enabled explicitly when the rest of the system is ready.

BUG=b:390242784
TEST=pin shows up in gpioget:
    ec:~> gpioget
    ...
      0  c_ec_pdb_ctrl

Change-Id: I3a70d220a974c9c235ce849f3146a47351e1a76e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6461503
Auto-Submit: caveh jalali <caveh@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: caveh jalali <caveh@chromium.org>
Commit-Queue: caveh jalali <caveh@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>